elements: ship JSX augmentations via elements.ts side-effect import
Restore main's elements.d.ts layout (now as elements.ts so tsc emits it
to dist/) and pull it in from index.ts via a side-effect import so
consumers actually receive the JSX augmentations for patchwork-view and
the new patchwork-component intrinsic.
